requirements.txt
flask
sentence-transformers
faiss-cpu
scikit-learn
requests
flask-cors
gunicorn
transformers
torch
